Structure tests directory to match implementation files
